[finance_report_ui] fix(#1868): S5 PR-C — split WorkflowNotifications god-file, fetch ratchet, ConfirmDialog input object, stage2 filters
- Split the 769-line components/workflow/WorkflowNotifications.tsx into one
  file per component/hook (WorkflowEventCard, WorkflowInbox,
  WorkflowNotificationCenter, WorkflowStatusFeed, UploadToReportHome,
  WorkflowEventsPageContent, hooks/useWorkflowStatus.ts, workflowFormat.ts
  helpers); original path kept as a thin barrel re-export so all 6 external
  consumers resolve unchanged. Export `cx` once from components/ui instead
  of re-implementing it locally.
- lib/reportPackage.ts extracts isValidReportDate/reportPeriodStart/
  packageSnapshotRequest/packageQuery out of the hook; deletes the
  duplicate packageReadinessQuery in reports/page.tsx.
- Migrate useLlmConfigStatus to useApiQuery instead of hand-rolled
  apiFetch + useState/useEffect orchestration.
- Install a shrink-only ratchet (common/testing/fe_fetch_ratchet.py) on
  direct apiFetch/apiStream/apiDelete/apiUpload call sites under
  components/+app/, generic-aware and seeded from the measured
  post-migration count (87).
- ConfirmDialog: showInput/inputLabel/inputPlaceholder/inputRequired
  collapse into a single input?: {label; placeholder?; required?} prop.
- Stage2Filters/RunSummaryPanel: individual value+setter props collapse
  into filters/onChange(patch) and unresolvedCounts/approval objects.
- New tooling gates: test_fe_no_godfile.py (no file under
  components/workflow/ exceeds 300 lines; cx has exactly one definition)
  and test_fe_fetch_ratchet.py, both wired to new ACs in
  common/meta/contract.py (AC-meta.fe-contract-types.5) and
  common/testing/contract.py (AC-testing.fe-fetch.1).
