+++ apps/frontend/src/components/review/stage2/Stage2Filters.tsx
@@ -1,26 +1,25 @@
 import { InfoHint } from "@/components/ui/InfoHint";
 
-interface Stage2FiltersProps {
-    checkTypeFilter: string;
-    statusFilter: string;
-    severityFilter: string[];
+export interface Stage2FiltersState {
+    checkType: string;
+    status: string;
+    severity: string[];
     minScore: number;
-    onToggleSeverity: (severity: string) => void;
-    onCheckTypeChange: (value: string) => void;
-    onStatusChange: (value: string) => void;
-    onMinScoreChange: (value: number) => void;
 }
 
-export function Stage2Filters({
-    checkTypeFilter,
-    statusFilter,
-    severityFilter,
-    minScore,
-    onToggleSeverity,
-    onCheckTypeChange,
-    onStatusChange,
-    onMinScoreChange,
-}: Stage2FiltersProps) {
+interface Stage2FiltersProps {
+    filters: Stage2FiltersState;
+    onChange: (patch: Partial<Stage2FiltersState>) => void;
+}
+
+export function Stage2Filters({ filters, onChange }: Stage2FiltersProps) {
+    const toggleSeverity = (severity: string) => {
+        const next = filters.severity.includes(severity)
+            ? filters.severity.filter((s) => s !== severity)
+            : [...filters.severity, severity];
+        onChange({ severity: next });
+    };
+
     return (
         <div className="mb-6 grid grid-cols-1 md:grid-cols-4 gap-4 bg-[var(--background-card)] p-4 rounded-lg border border-[var(--border)]">
             <div className="space-y-2">
@@ -30,9 +29,9 @@ export function Stage2Filters({
                         <button
                             key={s}
                             type="button"
-                            onClick={() => onToggleSeverity(s)}
+                            onClick={() => toggleSeverity(s)}
                             className={`px-2 py-1 text-xs rounded-full border transition-colors ${
-                                severityFilter.includes(s)
+                                filters.severity.includes(s)
                                     ? "bg-[var(--accent)] text-white border-[var(--accent)]"
                                     : "bg-[var(--background)] text-muted border-[var(--border)] hover:border-[var(--accent)]"
                             }`}
@@ -47,8 +46,8 @@ export function Stage2Filters({
                 <label className="text-xs font-medium text-muted uppercase">Check Type</label>
                 <select
                     className="input text-sm py-1"
-                    value={checkTypeFilter}
-                    onChange={(e) => onCheckTypeChange(e.target.value)}
+                    value={filters.checkType}
+                    onChange={(e) => onChange({ checkType: e.target.value })}
                 >
                     <option value="">All Types</option>
                     <option value="duplicate">Duplicate</option>
@@ -61,8 +60,8 @@ export function Stage2Filters({
                 <label className="text-xs font-medium text-muted uppercase">Status</label>
                 <select
                     className="input text-sm py-1"
-                    value={statusFilter}
-                    onChange={(e) => onStatusChange(e.target.value)}
+                    value={filters.status}
+                    onChange={(e) => onChange({ status: e.target.value })}
                 >
                     <option value="">All Statuses</option>
                     <option value="pending">Pending</option>
@@ -72,7 +71,7 @@ export function Stage2Filters({
 
             <div className="space-y-2">
                 <label className="text-xs font-medium text-muted uppercase">
-                    Min Match Score: {minScore}
+                    Min Match Score: {filters.minScore}
                     <InfoHint term="match_score" label="Match score" />
                 </label>
                 <input
@@ -80,8 +79,8 @@ export function Stage2Filters({
                     min="0"
                     max="100"
                     step="5"
-                    value={minScore}
-                    onChange={(e) => onMinScoreChange(parseInt(e.target.value))}
+                    value={filters.minScore}
+                    onChange={(e) => onChange({ minScore: parseInt(e.target.value) })}
                     className="w-full h-2 bg-[var(--border)] rounded-lg appearance-none cursor-pointer accent-[var(--accent)]"
                 />
             </div>
