+++ apps/frontend/src/hooks/useLlmConfigStatus.ts
@@ -1,9 +1,8 @@
 "use client";
 
-import { useCallback, useEffect, useState } from "react";
-
-import { fetchLlmConfigStatus } from "@/lib/api";
+import { useApiQuery } from "@/hooks/useApiQuery";
 import { getUserId } from "@/lib/auth";
+import type { LlmConfigStatusResponse } from "@/lib/types";
 
 export interface UseLlmConfigStatus {
   /** Whether the user has a usable LLM configuration; null until first load. */
@@ -17,32 +16,26 @@ export interface UseLlmConfigStatus {
 /**
  * LLM first-run status (EPIC-023 PR4).
  *
- * Fetches `GET /api/llm/config/status` on mount and exposes the configured
- * flag plus a manual `refresh`. Only runs when a local session id is present,
- * mirroring `useSessionBootstrap` — the status endpoint is authenticated and
- * the login route renders outside the authenticated shell.
+ * `GET /api/llm/config/status` via `useApiQuery` (#1868 S5 PR-C — was the
+ * pre-react-query useState/useEffect idiom). Only runs when a local session
+ * id is present, mirroring `useSessionBootstrap` — the status endpoint is
+ * authenticated and the login route renders outside the authenticated shell.
+ * A failed fetch leaves `configured` at its last-known value (react-query's
+ * `data` is untouched by a query error) rather than forcing the first-run
+ * modal on a transient error.
  */
 export function useLlmConfigStatus(): UseLlmConfigStatus {
-  const [configured, setConfigured] = useState<boolean | null>(null);
-  const [loading, setLoading] = useState(false);
+  const query = useApiQuery<LlmConfigStatusResponse>(
+    ["llm", "config-status"],
+    "/api/llm/config/status",
+    { enabled: getUserId() !== null },
+  );
 
-  const refresh = useCallback(async () => {
-    if (getUserId() === null) return;
-    setLoading(true);
-    try {
-      const status = await fetchLlmConfigStatus();
-      setConfigured(status.configured);
-    } catch {
-      // A 401 redirects via apiFetch. For any other failure, leave the flag
-      // untouched rather than forcing the first-run modal on a transient error.
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    void refresh();
-  }, [refresh]);
-
-  return { configured, loading, refresh };
+  return {
+    configured: query.data?.configured ?? null,
+    loading: query.isFetching,
+    refresh: async () => {
+      await query.refetch();
+    },
+  };
 }
