+++ apps/frontend/src/hooks/usePersonalReportPackage.ts
@@ -2,6 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 
 import { apiFetch } from "@/lib/api";
 import { useApiQuery } from "@/hooks/useApiQuery";
+import { isValidReportDate, packageQuery, packageSnapshotRequest } from "@/lib/reportPackage";
 import type {
   AnnualizedIncomeScheduleResponse,
   FrameworkPolicyResult,
@@ -12,50 +13,6 @@ import type {
   PersonalReportPackageSnapshotSummary,
   PersonalReportPackageTraceabilityResponse,
 } from "@/lib/types";
-
-const REPORT_DATE_PATTERN = /^\d{4}-\d{2}-\d{2}$/;
-
-export function isValidReportDate(reportDate: string): boolean {
-  if (!REPORT_DATE_PATTERN.test(reportDate)) return false;
-  const [year, month, day] = reportDate.split("-").map(Number);
-  const parsed = new Date(Date.UTC(year, month - 1, day));
-  return (
-    parsed.getUTCFullYear() === year &&
-    parsed.getUTCMonth() === month - 1 &&
-    parsed.getUTCDate() === day
-  );
-}
-
-export function reportPeriodStart(reportDate: string): string {
-  const [year, month, day] = reportDate.split("-").map(Number);
-  if (!year || !month || !day) return reportDate;
-  const previousYear = year - 1;
-  const lastDayOfMonth = new Date(Date.UTC(previousYear, month, 0)).getUTCDate();
-  const clampedDay = Math.min(day, lastDayOfMonth);
-  return `${previousYear}-${String(month).padStart(2, "0")}-${String(clampedDay).padStart(2, "0")}`;
-}
-
-export function packageSnapshotRequest(
-  frameworkId: string,
-  reportDate: string,
-  currency = "SGD",
-) {
-  return {
-    framework_id: frameworkId,
-    start_date: reportPeriodStart(reportDate),
-    end_date: reportDate,
-    as_of_date: reportDate,
-    currency,
-  };
-}
-
-function packageQuery(reportDate: string, frameworkId?: string): string {
-  const params = new URLSearchParams(frameworkId ? { framework_id: frameworkId } : undefined);
-  params.set("start_date", reportPeriodStart(reportDate));
-  params.set("end_date", reportDate);
-  params.set("as_of_date", reportDate);
-  return `?${params.toString()}`;
-}
 
 export async function generatePackageSnapshot(
   frameworkId: string,
