+++ common/testing/fe-fetch-ratchet-baseline.json
@@ -1,0 +1,3 @@
+{
+  "total": 87
+}
